+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - Requesting a single file through the JSON API now works as well
     - This allows the nsrr gem to download single files
   - Dataset editors can now view and filter agreements for single datasets
+  - The dataset files directory is now created on dataset creation
 - **Dashboard Changes**
   - Improved display of admin pages linked from the dashboard
 - **Forum Changes**